修改gto奖励

+++ src/main/java/com/longti/upjc/entity/sporttery/LOTO_ORDER.java
@@ -22,7 +22,7 @@ public class LOTO_ORDER implements Serializable {
     private String issue; //期号
     private String vsteam; //对阵信息
     private String vsresult; //对阵结果
-    private Integer bet_fee; //投注金额
+    private Long bet_fee; //投注金额
     private Integer win_fee; //奖金,0表示未中奖
     private Date create_time; //创建时间
     private Date start_create_time; //开始创建时间
@@ -86,10 +86,10 @@ public class LOTO_ORDER implements Serializable {
 	public void setVsresult(String vsresult) {
 		this.vsresult = vsresult;
 	}
-    public Integer getBet_fee() {
+    public Long getBet_fee() {
         return bet_fee;
 	}
-	public void setBet_fee(Integer bet_fee) {
+	public void setBet_fee(Long bet_fee) {
 		this.bet_fee = bet_fee;
 	}
     public Integer getWin_fee() {
